+++ apps/backend/src/modules/user-routes/libs/types/types.ts
@@ -1,8 +1,8 @@
-export { type UserRouteFilter } from "./user-route.filter.type.js";
+export { type UserRouteGetAllFilters } from "./user-route-get-all-filters.type.js";
 export {
 	type UserRouteDeleteParameters,
 	type UserRouteFinishRequestDto,
-	type UserRouteGetAllOptions,
+	type UserRouteGetAllQueryRequestDto,
 	type UserRouteQueryRequestDto,
 	type UserRouteResponseDto,
 	type UserRouteStatusType,