+++ apps/backend/src/modules/user-routes/libs/types/user-route.filter.type.ts
@@ -1,5 +1,0 @@
-import { type UserRouteEntity } from "../../user-route.entity.js";
-
-type UserRouteFilter = Partial<ReturnType<UserRouteEntity["toObject"]>>;
-
-export { type UserRouteFilter };