+++ apps/backend/src/modules/user-routes/user-route.entity.ts
@@ -1,19 +1,23 @@
 import { type Entity, type LineStringGeometry } from "~/libs/types/types.js";
 
-import { type UserRouteStatusType } from "./libs/types/type.js";
+import { type UserRouteStatusType } from "./libs/types/types.js";
 
 class UserRouteEntity implements Entity {
 	private actualGeometry: LineStringGeometry;
 
 	private completedAt: null | string;
 
+	private distance: number;
+
 	private id: null | number;
 
 	private plannedGeometry: LineStringGeometry;
 
+	private reviewComment: null | string;
+
 	private routeId: number;
 
-	private routeName: string;
+	private routeName: null | string;
 
 	private startedAt: null | string;
 
@@ -24,8 +28,10 @@ class UserRouteEntity implements Entity {
 	private constructor({
 		actualGeometry,
 		completedAt,
+		distance,
 		id,
 		plannedGeometry,
+		reviewComment,
 		routeId,
 		routeName,
 		startedAt,
@@ -34,10 +40,12 @@ class UserRouteEntity implements Entity {
 	}: {
 		actualGeometry: LineStringGeometry;
 		completedAt: null | string;
+		distance: number;
 		id: null | number;
 		plannedGeometry: LineStringGeometry;
+		reviewComment?: null | string;
 		routeId: number;
-		routeName: string;
+		routeName?: null | string;
 		startedAt: null | string;
 		status: null | UserRouteStatusType;
 		userId: number;
@@ -47,17 +55,21 @@ class UserRouteEntity implements Entity {
 		this.completedAt = completedAt;
 		this.plannedGeometry = plannedGeometry;
 		this.routeId = routeId;
+		this.routeName = routeName ?? null;
+		this.reviewComment = reviewComment ?? null;
+		this.distance = distance;
 		this.startedAt = startedAt;
 		this.status = status;
 		this.userId = userId;
-		this.routeName = routeName;
 	}
 
 	public static initialize(data: {
 		actualGeometry: LineStringGeometry;
 		completedAt: null | string;
+		distance: number;
 		id: number;
 		plannedGeometry: LineStringGeometry;
+		reviewComment?: null | string;
 		routeId: number;
 		routeName: string;
 		startedAt: null | string;
@@ -67,8 +79,10 @@ class UserRouteEntity implements Entity {
 		return new UserRouteEntity({
 			actualGeometry: data.actualGeometry,
 			completedAt: data.completedAt,
+			distance: data.distance,
 			id: data.id,
 			plannedGeometry: data.plannedGeometry,
+			reviewComment: data.reviewComment ?? null,
 			routeId: data.routeId,
 			routeName: data.routeName,
 			startedAt: data.startedAt,
@@ -79,26 +93,26 @@ class UserRouteEntity implements Entity {
 
 	public static initializeNew({
 		actualGeometry,
+		distance,
 		plannedGeometry,
 		routeId,
-		routeName,
 		status,
 		userId,
 	}: {
 		actualGeometry: LineStringGeometry;
+		distance: number;
 		plannedGeometry: LineStringGeometry;
 		routeId: number;
-		routeName: string;
 		status: UserRouteStatusType;
 		userId: number;
 	}): UserRouteEntity {
 		return new UserRouteEntity({
 			actualGeometry,
 			completedAt: null,
+			distance,
 			id: null,
 			plannedGeometry,
 			routeId,
-			routeName,
 			startedAt: null,
 			status,
 			userId,
@@ -128,8 +142,10 @@ class UserRouteEntity implements Entity {
 	public toObject(): {
 		actualGeometry: LineStringGeometry;
 		completedAt: null | string;
+		distance: number;
 		id: number;
 		plannedGeometry: LineStringGeometry;
+		reviewComment: null | string;
 		routeId: number;
 		routeName: string;
 		startedAt: null | string;
@@ -139,10 +155,12 @@ class UserRouteEntity implements Entity {
 		return {
 			actualGeometry: this.actualGeometry,
 			completedAt: this.completedAt,
+			distance: this.distance,
 			id: this.id as number,
 			plannedGeometry: this.plannedGeometry,
+			reviewComment: this.reviewComment,
 			routeId: this.routeId,
-			routeName: this.routeName,
+			routeName: this.routeName as string,
 			startedAt: this.startedAt,
 			status: this.status as UserRouteStatusType,
 			userId: this.userId,