+++ apps/frontend/src/libs/components/carousel/libs/hooks/hooks.ts
@@ -1,4 +1,0 @@
-export { useCarouselMomentum } from "./use-carousel-momentum.hook.js";
-export { useMouseLeave } from "./use-carousel-mouse-leave.hook.js";
-export { useCarouselReference } from "./use-carousel-reference.hook.js";
-export { useCarouselWheelEvent } from "./use-carousel-wheel-event.hook.js";