+++ apps/frontend/src/libs/components/carousel/libs/types/types.ts
@@ -1,1 +1,0 @@
-export { type CarouselReference } from "./carousel-reference.type.js";