+++ apps/frontend/src/libs/components/components.ts
@@ -2,7 +2,6 @@ export { AppLayout } from "./app-layout/app-layout.js";
 export { App } from "./app/app.js";
 export { Avatar } from "./avatar/avatar.js";
 export { Button } from "./button/button.js";
-export { Carousel } from "./carousel/carousel.js";
 export { Checkbox } from "./checkbox/checkbox.js";
 export { CreatePOIModal } from "./create-poi-modal/create-poi-modal.js";
 export { CreateReviewModal } from "./create-review-modal/create-review-modal.js";