+++ apps/frontend/src/pages/dashboard/libs/components/card/dashboard-route-card.tsx
@@ -9,18 +9,22 @@ type Properties = {
 	route: UserRouteResponseDto;
 };
 
-const UserHistoryCard = ({ route }: Properties): React.JSX.Element => {
+const DashboardRouteCard = ({ route }: Properties): React.JSX.Element => {
 	const routeLine = { geometry: route.actualGeometry, id: route.routeName };
 
 	return (
 		<Link
 			className={styles["card"]}
-			to={configureString(AppRoute.ROUTES_$ID, { id: route.id.toString() })}
+			to={configureString(AppRoute.ROUTES_$ID, {
+				id: route.routeId.toString(),
+			})}
 		>
+			<div className={styles["map-container"]}>
+				<MapProvider routeLine={routeLine} />
+			</div>
 			<h2 className={styles["label"]}>{route.routeName}</h2>
-			<MapProvider routeLine={routeLine} />
 		</Link>
 	);
 };
 
-export { UserHistoryCard };
+export { DashboardRouteCard };