+++ packages/shared/src/modules/user-routes/libs/types/user-route-response-dto.type.ts
@@ -4,8 +4,10 @@ import { type UserRouteStatusType } from "./user-route-status.type.js";
 type UserRouteResponseDto = {
 	actualGeometry: LineStringGeometry;
 	completedAt: null | string;
+	distance: number;
 	id: number;
 	plannedGeometry: LineStringGeometry;
+	reviewComment: null | string;
 	routeId: number;
 	routeName: string;
 	startedAt: null | string;