+++ packages/shared/src/modules/user-routes/validation-schemas/validation-schemas.ts
@@ -1,4 +1,5 @@
 export { userRouteDeleteParameters } from "./user-route-delete.validation-schema.js";
 export { userRouteFinish } from "./user-route-finish.validation-schema.js";
+export { userRouteGetAll } from "./user-route-get-all.validation-schema.js";
 export { userRouteQuery } from "./user-route-query.validation-schema.js";
 export { userRouteStart } from "./user-route-start.validation-schema.js";